kernvind/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["bsinghvt <baljotsingh1986@yahoo.com>"]
readme = "README.md"
packages = [{include = "app", from = "src"}]

[tool.poetry.scripts]
start_dev = "manage:dev_app"
start_dev_db_schemas = "manage:dev_app_generate_dbschemas"

[tool.poetry.dependencies]
python = "^3.12"
quart = {extras = ["dotenv"], version = "^0.19.6"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.6"}
quart-bcrypt = "^0.0.9"
email-validator = "^2.2.0"
quart-jwt-extended = {extras = ["asymmetric-crypto"], version = "^0.1.0"}
crypto-lib = {path = "../../crypto_lib"}
quart-cors = "^0.7.0"
google-auth = "^2.35.0"
google-api-python-client = "^2.149.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
langchain-postgres = "^0.0.12"
psycopg = {extras = ["binary"], version = "^3.2.2"}
langchain-huggingface = "^0.1.1"
langchain-ollama = "^0.2.0"
langchain-openai = "^0.2.5"
data-models = {path = "../../data_models"}
quart-schema = "^0.20.0"
pydantic = "^2.10.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
